removed src and include??

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: clouden <clouden@student.42madrid.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/30 11:28:05 by clouden           #+#    #+#             */
-/*   Updated: 2025/05/07 17:50:32 by clouden          ###   ########.fr       */
+/*   Updated: 2025/05/15 20:33:21 by clouden          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 void *ft_calloc(size_t count, size_t size)
 {
     void *mem;
-    int i;
+    size_t i;
 
     i = 0;
     if (count == 0 || size == 0)
