+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: clouden <clouden@student.42madrid.com      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/17 11:59:11 by clouden           #+#    #+#             */
-/*   Updated: 2025/05/14 23:50:45 by clouden          ###   ########.fr       */
+/*   Updated: 2025/05/18 11:05:58 by clouden          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 size_t ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int i; 
+	size_t i; 
 	
 	i = 0;
 	if (size > 0)
